requirements.txt
gitdb
gitpython
tornado==6.1.0
markdown
watchdog
